pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "A travel agent that can help you find places to eat, drink or see."
authors = [
    {name = "Christoph Pernul", email = "cpernul@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.85.0",
    "openai-agents>=0.0.17",
    "gradio>=5.33.1",
    "httpx>=0.28.1",
    "bs4>=0.0.2",
    "requests>=2.32.4",
    "python-dotenv>=1.1.0",
    "setuptools>=80.9.0",
]



[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
]
